clearly-defined: Split out "curation(s)" and "definition" data classes
Move the data classes for [1], [2] and [3] to separate files for a better
overview.

[1]: https://github.com/clearlydefined/service/blob/4917725/schemas/curation-1.0.json
[2]: https://github.com/clearlydefined/service/blob/b339cb7/schemas/curations-1.0.json
[3]: https://github.com/clearlydefined/service/blob/b339cb7/schemas/definition-1.0.json
