+++ clients/clearly-defined/src/main/kotlin/ClearlyDefinedService.kt
@@ -19,18 +19,12 @@
 
 package org.ossreviewtoolkit.clients.clearlydefined
 
-import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonValue
 import com.fasterxml.jackson.databind.DeserializationFeature
-import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.json.JsonMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
-
-import java.io.File
-import java.net.URI
 
 import okhttp3.OkHttpClient
 import okhttp3.ResponseBody
@@ -133,179 +127,6 @@ interface ClearlyDefinedService {
     }
 
     /**
-     * See https://github.com/clearlydefined/service/blob/b339cb7/schemas/definition-1.0.json#L80-L89.
-     */
-    data class FinalScore(
-        val effective: Int,
-        val tool: Int
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/b339cb7/schemas/definition-1.0.json#L48-L61.
-     */
-    data class Meta(
-        val schemaVersion: String,
-        val updated: String
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/curation-1.0.json#L7-L17.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Curation(
-        val described: Described? = null,
-        val licensed: Licensed? = null,
-        val files: List<FileEntry>? = null
-    )
-
-    data class ContributedCurations(
-        val curations: Map<Coordinates, Curation>,
-        val contributions: List<JsonNode>
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L145-L179.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Described(
-        val score: DescribedScore? = null,
-        val toolScore: DescribedScore? = null,
-        val facets: Facets? = null,
-        val sourceLocation: SourceLocation? = null,
-        val urls: URLs? = null,
-        val projectWebsite: URI? = null,
-        val issueTracker: URI? = null,
-        val releaseDate: String? = null,
-        val hashes: Hashes? = null,
-        val files: Int? = null,
-        val tools: List<String>? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L180-L190.
-     */
-    data class DescribedScore(
-        val total: Int,
-        val date: Int,
-        val source: Int
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L264-L275.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Facets(
-        val core: Facet? = null,
-        val data: Facet? = null,
-        val dev: Facet? = null,
-        val doc: Facet? = null,
-        val examples: Facet? = null,
-        val tests: Facet? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L276-L286.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Facet(
-        val files: Int? = null,
-        val attribution: Attribution? = null,
-        val discovered: Discovered? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L287-L301.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Attribution(
-        val parties: List<String>? = null,
-        val unknown: Int? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L305-L319.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Discovered(
-        val expressions: List<String>? = null,
-        val unknown: Int? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L211-L235.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class SourceLocation(
-        // The following properties match those of Coordinates, except that the revision is mandatory here.
-        val type: ComponentType,
-        val provider: Provider,
-        val namespace: String? = null,
-        val name: String,
-        val revision: String,
-
-        val path: String? = null,
-        val url: String? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L236-L253.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class URLs(
-        val registry: URI? = null,
-        val version: URI? = null,
-        val download: URI? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L135-L144.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Hashes(
-        val md5: String? = null,
-        val sha1: String? = null,
-        val sha256: String? = null,
-        val gitSha: String? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L90-L134.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class FileEntry(
-        val path: File,
-        val license: String? = null,
-        val attributions: List<String>? = null,
-        val facets: Facets? = null,
-        val hashes: Hashes? = null,
-        val token: String? = null,
-        val natures: Set<Nature>? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L254-L263.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Licensed(
-        val score: LicensedScore? = null,
-        val toolScore: LicensedScore? = null,
-        val declared: String? = null,
-        val facets: Facets? = null
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/4917725/schemas/definition-1.0.json#L191-L204.
-     */
-    data class LicensedScore(
-        val total: Int,
-        val declared: Int,
-        val discovered: Int,
-        val consistency: Int,
-        val spdx: Int,
-        val texts: Int
-    )
-
-    /**
      * See https://github.com/clearlydefined/service/blob/4e210d7/schemas/swagger.yaml#L84-L101.
      */
     data class ContributionPatch(
@@ -340,67 +161,6 @@ interface ClearlyDefinedService {
          */
         val removedDefinitions: Boolean
     )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/b339cb7/schemas/curations-1.0.json#L8-L15.
-     */
-    data class Patch(
-        val coordinates: Coordinates,
-        val revisions: Map<String, Curation>
-    )
-
-    /**
-     * See https://github.com/clearlydefined/service/blob/b339cb7/schemas/curations-1.0.json#L64-L83 and
-     * https://docs.clearlydefined.io/using-data#a-note-on-definition-coordinates.
-     */
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    data class Coordinates(
-        /**
-         * The type of the component. For example, npm, git, nuget, maven, etc. This talks about the shape of the
-         * component.
-         */
-        val type: ComponentType,
-
-        /**
-         * Where the component can be found. Examples include npmjs, mavencentral, github, nuget, etc.
-         */
-        val provider: Provider,
-
-        /**
-         * Many component systems have namespaces: GitHub orgs, NPM namespace, Maven group id, etc. This segment must be
-         * supplied. If your component does not have a namespace, use '-' (ASCII hyphen).
-         */
-        val namespace: String? = null,
-
-        /**
-         * The name of the component. Given the mentioned [namespace] segment, this is just the simple name.
-         */
-        val name: String,
-
-        /**
-         * Components typically have some differentiator like a version or commit id. Use that here. If this segment is
-         * omitted, the latest revision is used (if that makes sense for the provider).
-         */
-        val revision: String? = null
-    ) {
-        companion object {
-            @JsonCreator
-            @JvmStatic
-            fun fromString(value: String): Coordinates {
-                val parts = value.split('/', limit = 5)
-                return Coordinates(
-                    type = ComponentType.fromString(parts[0]),
-                    provider = Provider.fromString(parts[1]),
-                    namespace = parts[2].takeUnless { it == "-" },
-                    name = parts[3],
-                    revision = parts.getOrNull(4)
-                )
-            }
-        }
-
-        @JsonValue
-        override fun toString() = listOfNotNull(type, provider, namespace ?: "-", name, revision).joinToString("/")
-    }
 
     /**
      * See https://github.com/clearlydefined/service/blob/53acc01/routes/curations.js#L86-L89.
