+++ scanner/src/main/kotlin/storages/ClearlyDefinedStorage.kt
@@ -29,6 +29,8 @@ import kotlinx.coroutines.runBlocking
 
 import org.ossreviewtoolkit.clients.clearlydefined.ClearlyDefinedService
 import org.ossreviewtoolkit.clients.clearlydefined.ComponentType
+import org.ossreviewtoolkit.clients.clearlydefined.Coordinates
+import org.ossreviewtoolkit.clients.clearlydefined.SourceLocation
 import org.ossreviewtoolkit.model.ArtifactProvenance
 import org.ossreviewtoolkit.model.Hash
 import org.ossreviewtoolkit.model.Identifier
@@ -60,10 +62,9 @@ import retrofit2.HttpException
 private const val TOOL_SCAN_CODE = "scancode"
 
 /**
- * Convert a [ClearlyDefinedService.SourceLocation] to a [ClearlyDefinedService.Coordinates] object.
+ * Convert a [SourceLocation] to a [Coordinates] object.
  */
-private fun ClearlyDefinedService.SourceLocation.toCoordinates(): ClearlyDefinedService.Coordinates =
-    ClearlyDefinedService.Coordinates(type, provider, namespace, name, revision)
+private fun SourceLocation.toCoordinates(): Coordinates = Coordinates(type, provider, namespace, name, revision)
 
 /**
  * Convert a [VcsInfo] to a [VcsInfoCurationData] object.
@@ -79,7 +80,7 @@ private fun packageCoordinates(
     id: Identifier,
     vcs: VcsInfo?,
     sourceArtifact: RemoteArtifact?
-): ClearlyDefinedService.Coordinates {
+): Coordinates {
     val sourceLocation = id.toClearlyDefinedSourceLocation(vcs?.toVcsInfoCurationData(), sourceArtifact)
     return sourceLocation?.toCoordinates() ?: id.toClearlyDefinedCoordinates()
         ?: throw IllegalArgumentException("Unable to create ClearlyDefined coordinates for '${id.toCoordinates()}'.")
@@ -89,7 +90,7 @@ private fun packageCoordinates(
  * Given a list of [tools], return the version of ScanCode that was used to scan the package with the given
  * [coordinates], or return null if no such tool entry is found.
  */
-private fun findScanCodeVersion(tools: List<String>, coordinates: ClearlyDefinedService.Coordinates): String? {
+private fun findScanCodeVersion(tools: List<String>, coordinates: Coordinates): String? {
     val toolUrl = "$coordinates/$TOOL_SCAN_CODE/"
     return tools.find { it.startsWith(toolUrl) }?.substring(toolUrl.length)
 }
@@ -143,10 +144,7 @@ class ClearlyDefinedStorage(
      * using the given [coordinates] for looking up the data. These may not necessarily be equivalent to the
      * identifier, as we try to lookup source code results if a GitHub repository is known.
      */
-    private suspend fun readFromClearlyDefined(
-        id: Identifier,
-        coordinates: ClearlyDefinedService.Coordinates
-    ): Result<List<ScanResult>> {
+    private suspend fun readFromClearlyDefined(id: Identifier, coordinates: Coordinates): Result<List<ScanResult>> {
         val startTime = Instant.now()
         log.info { "Looking up results for '${id.toCoordinates()}'." }
 
@@ -193,7 +191,7 @@ class ClearlyDefinedStorage(
      * The results have been produced by ScanCode in the given [version]; use the [startTime] for metadata.
      */
     private suspend fun loadScanCodeResults(
-        coordinates: ClearlyDefinedService.Coordinates,
+        coordinates: Coordinates,
         version: String,
         startTime: Instant
     ): Result<List<ScanResult>> {
